Fixed a reader bug where values were being sign-extended right rather than shifted right. Fixed a writer bug where in an edge condition the bits for a patch are reduced, while the bits for the value are increased.  The increased value was not stored in the output stream.

+++ src/ApacheOrcDotNet/Encodings/IntegerRunLengthEncodingV2Reader.cs
@@ -135,14 +135,14 @@ namespace ApacheOrcDotNet.Encodings
 		{
 			var raw = patchListValues[patchIndex];
 			patchIndex++;
-			long curGap = raw >> patchWidth;
+			long curGap = (long)((ulong)raw >> patchWidth);
 			patch = raw & patchMask;
 			while (curGap == 255 && patch == 0)
 			{
 				gap += 255;
 				raw = patchListValues[patchIndex];
 				patchIndex++;
-				curGap = raw >> patchWidth;
+				curGap = (long)((ulong)raw >> patchWidth);
 				patch = raw & patchMask;
 			}
 			gap += curGap;
